Remove several references to v0 from v1 package
