+++ scalafix-core/src/main/scala/scalafix/internal/rule/Disable.scala
@@ -4,7 +4,7 @@ import metaconfig.{Conf, Configured}
 import scala.meta._
 import scala.meta.transversers.Traverser
 import scalafix.internal.config.{DisableConfig, DisabledSymbol}
-import scalafix.internal.patch.DocSemanticdbIndex.InputSynthetic
+import scalafix.internal.v0.InputSynthetic
 import scalafix.internal.util.SymbolOps
 import scalafix.lint.{LintCategory, LintMessage}
 import scalafix.v0._
