+++ scalafix-core/src/main/scala/scalafix/internal/v0/DeprecatedRuleCtx.scala
@@ -1,14 +1,15 @@
-package scalafix.internal.patch
+package scalafix.internal.v0
 
 import org.scalameta.FileLine
 import scala.meta.Input
 import scala.meta.Tree
 import scala.meta.contrib.AssociatedComments
 import scala.meta.tokens.Tokens
-import scalafix.v0.RuleCtx
+import scalafix.internal.patch.DeprecatedPatchOps
 import scalafix.util.MatchingParens
 import scalafix.util.SemanticdbIndex
 import scalafix.util.TokenList
+import scalafix.v0.RuleCtx
 import scalafix.v1.Doc
 
 class DeprecatedRuleCtx(doc: Doc) extends RuleCtx with DeprecatedPatchOps {
