+++ scalafix-core/src/main/scala/scalafix/internal/v0/LegacyCodePrinter.scala
@@ -1,13 +1,11 @@
 package scalafix.internal.v0
 
 import java.nio.charset.StandardCharsets
-import scala.meta._
 import scala.meta.inputs.Input
 import scala.meta.inputs.Position
 import scala.meta.internal.ScalametaInternals
 import scala.meta.internal.semanticdb.Scala._
 import scala.meta.internal.{semanticdb => s}
-import scalafix.internal.patch.DocSemanticdbIndex.InputSynthetic
 import scalafix.v0
 import scalafix.v0.ResolvedName
 import scalafix.v1
