+++ scalafix-core/src/main/scala/scalafix/internal/v1/Rules.scala
@@ -8,6 +8,8 @@ import scalafix.internal.config.MetaconfigPendingUpstream
 import scalafix.internal.config.NoInferConfig
 import scalafix.internal.rule._
 import scalafix.internal.util.SuppressOps
+import scalafix.internal.v0.LegacySemanticRule
+import scalafix.internal.v0.LegacySyntacticRule
 import scalafix.lint.LintDiagnostic
 import scalafix.lint.LintMessage
 import scalafix.patch.Patch
@@ -57,11 +59,7 @@ case class Rules(rules: List[Rule] = Nil) {
       case rule: SyntacticRule =>
         rule.name -> rule.fix(doc.doc)
     }.toMap
-    scalafix.Patch.apply(
-      fixes,
-      doc.doc.toRuleCtx,
-      Some(doc.toSemanticdbIndex),
-      suppress)
+    scalafix.Patch.semantic(fixes, doc, suppress)
   }
 
   def syntacticPatch(
@@ -71,7 +69,7 @@ case class Rules(rules: List[Rule] = Nil) {
     val fixes = syntacticRules.iterator.map { rule =>
       rule.name -> rule.fix(doc)
     }.toMap
-    scalafix.Patch.apply(fixes, doc.toRuleCtx, None, suppress)
+    scalafix.Patch.syntactic(fixes, doc, suppress)
   }
 }
